% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_ranks}
\alias{all_ranks}
\title{Ranks handled by \code{yatah}}
\format{
An object of class \code{character} of length 8.
}
\usage{
all_ranks
}
\description{
Ranks handled by \code{yatah}
}
\examples{
all_ranks
}
\keyword{datasets}
