% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{last_clade}
\alias{last_rank}
\alias{all_clades}
\title{Deprecated functions}
\usage{
last_clade(...)

last_rank(...)

all_clades(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Use \code{\link[=get_last_clade]{get_last_clade()}} instead of \code{last_clade()}.

Use \code{\link[=get_last_rank]{get_last_rank()}} instead of \code{last_rank()}.

Use \code{\link[=get_all_clades]{get_all_clades()}} instead of \code{all_clades()}.
}
\keyword{internal}
