% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{depth}
\alias{depth}
\title{Common depth}
\usage{
depth(lineage)
}
\arguments{
\item{lineage}{string. Vector of lineages.}
}
\description{
Throw an error if depth is not the same across lineages.
}
\keyword{internal}
