% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\docType{data}
\name{.allchr}
\alias{.allchr}
\title{Characters allowed in lineages}
\format{
An object of class \code{character} of length 1.
}
\usage{
.allchr
}
\description{
Characters allowed in lineages
}
\keyword{datasets}
