% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_last_clade}
\alias{get_last_clade}
\title{Extract the last clade of a lineage}
\usage{
get_last_clade(lineage, same = TRUE)
}
\arguments{
\item{lineage}{string. Vector of lineages.}

\item{same}{logical. Does the lineage have the same depth? Default to \code{TRUE}.}
}
\value{
A string. The last clades of the given lineages.
}
\description{
Extract the last clade of a lineage
}
\examples{
lineage1 <- "k__Bacteria|p__Verrucomicrobia|c__Verrucomicrobiae"
lineage2 <- "k__Bacteria|p__Firmicutes|c__Clostridia"
get_last_clade(c(lineage1, lineage2))
}
