% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels.R
\name{is_level}
\alias{is_level}
\title{Test if a lineage go down to a specified level}
\usage{
is_level(lineage, level = c("kingdom", "phylum", "class", "order",
  "family", "genus", "species", "strain"))
}
\arguments{
\item{lineage}{string.}

\item{level}{string. One of \code{c("kingdom", "phylum", "class",
"order", "family", "genus", "species", "strain")} with partial matching.}
}
\value{
logical.
}
\description{
Test if a lineage go down to a specified level
}
\examples{
lineage <- "k__Bacteria|p__Verrucomicrobia|c__Verrucomicrobiae"
is_level(lineage, "class")
is_level(lineage, "genus")
}
