% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels.R
\name{last_level}
\alias{last_level}
\title{Extract the last level of a lineage}
\usage{
last_level(lineage, same = TRUE)
}
\arguments{
\item{lineage}{string.}

\item{same}{logical. Do the lineage must be at the same depth? Default to TRUE.}
}
\value{
A string.
}
\description{
Extract the last level of a lineage
}
\examples{
lineage1 <- "k__Bacteria|p__Verrucomicrobia|c__Verrucomicrobiae"
lineage2 <- "k__Bacteria|p__Firmicutes|c__Clostridia"
last_level(c(lineage1, lineage2))
}
