% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxtree.R
\name{taxtree}
\alias{taxtree}
\title{Taxonomic tree}
\usage{
taxtree(table, collapse = TRUE, lineage_length = 1, root = "")
}
\arguments{
\item{table}{dataframe.}

\item{collapse}{logical. Should node with one child be vanished? Default to TRUE.}

\item{lineage_length}{double. Mean lineage length. Default to 1.}

\item{root}{character. Name of the root if no natural root is given.}
}
\description{
Compute taxonomic tree from taxonomic table
}
