% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yatah-package.R
\docType{package}
\name{yatah-package}
\alias{yatah}
\alias{yatah-package}
\title{yatah: Yet Another TAxonomy Handler}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Provides functions to manage taxonomy when lineages are described with strings and ranks separated with special patterns like "|*__" or ";*__".
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/abichat/yatah}
  \item \url{https://abichat.github.io/yatah}
  \item Report bugs at \url{https://github.com/abichat/yatah/issues}
}

}
\author{
\strong{Maintainer}: Antoine Bichat \email{a.bichat@yahoo.fr} (\href{https://orcid.org/0000-0001-6599-7081}{ORCID})

}
\keyword{internal}
