% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundances.R
\docType{data}
\name{abundances}
\alias{abundances}
\title{Abundance table for 199 samples.}
\format{A data.frame with 1585 rows and 200 variables:
\describe{
\item{lineages}{lineage (string)}
\item{XXX}{abundance of each lineage in the sample XXX (double)}
}}
\source{
\href{https://www.embopress.org/doi/pdf/10.15252/msb.20145645}{Zeller et al., 2014}, \href{https://www.nature.com/articles/nmeth.4468}{Pasolli et al., 2017}
}
\usage{
abundances
}
\description{
A dataset containing the abundances of 1585 lineages among 199 patients.
}
\keyword{datasets}
