% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{error_lineage}
\alias{error_lineage}
\title{Throw error if the string is not a lineage}
\usage{
error_lineage(string, sep = "\\\\|")
}
\arguments{
\item{string}{string.}

\item{sep}{string. Rank separator. Default to \code{\\\\|} but
\code{;} could be used too.}
}
\description{
Throw error if the string is not a lineage
}
