% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_tests.R
\name{is_lineage}
\alias{is_lineage}
\title{Test if a string is a lineage}
\usage{
is_lineage(string, sep = NULL)
}
\arguments{
\item{string}{string to be tested as lineage.}

\item{sep}{string. Rank separator. Default to \code{|} (regex \code{\\\\|}) but
could be modified with \code{options(yatah_sep = ...)}.}
}
\value{
A logical.
}
\description{
Test if a string is a lineage
}
\examples{
is_lineage("k__Bacteria|p__Firmicutes|c__Clostridia|o__Clostridiales")
}
