% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_tests.R
\name{is_lineage}
\alias{is_lineage}
\title{Test if a string is a lineage}
\usage{
is_lineage(string)
}
\arguments{
\item{string}{string to be tested as lineage.}
}
\value{
A logical.
}
\description{
Test if a string is a lineage
}
\details{
Alphanumeric character, hyphen, dots, square brackets
and non-consecutive underscores are allowed in clades names.
}
\examples{
is_lineage("k__Bacteria|p__Firmicutes|c__Clostridia|o__Clostridiales")
}
