% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_tests.R
\name{is_lineage}
\alias{is_lineage}
\title{Test if a string is a lineage}
\usage{
is_lineage(string, sep = "\\\\|")
}
\arguments{
\item{string}{string.}

\item{sep}{string. Rank separator. Default to \code{\\\\|} but
\code{;} could be used too.}
}
\value{
A logical.
}
\description{
Test if a string is a lineage
}
\examples{
is_lineage("k__Bacteria|p__Firmicutes|c__Clostridia|o__Clostridiales")
}
