% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{last_rank}
\alias{last_rank}
\title{Extract the last rank of a lineage}
\usage{
last_rank(lineage, same = TRUE)
}
\arguments{
\item{lineage}{string. Vector of lineages.}

\item{same}{logical. Does the lineage have the same depth? Default to TRUE.}
}
\value{
A string. The last rank of the given lineages.
}
\description{
Extract the last rank of a lineage
}
\examples{
lineage1 <- "k__Bacteria|p__Verrucomicrobia|c__Verrucomicrobiae"
lineage2 <- "k__Bacteria|p__Firmicutes|c__Clostridia"
last_rank(c(lineage1, lineage2))
}
